LPS-67473 Use GetterUtil.getLong in case the layoutId is greater than the maximum integer size

+++ modules/apps/web-experience/layout/layout-admin-web/src/main/java/com/liferay/layout/admin/web/internal/exportimport/data/handler/LayoutStagedModelDataHandler.java
@@ -339,7 +339,7 @@ public class LayoutStagedModelDataHandler
 		Element layoutElement =
 			portletDataContext.getImportDataStagedModelElement(layout);
 
-		long layoutId = GetterUtil.getInteger(
+		long layoutId = GetterUtil.getLong(
 			layoutElement.attributeValue("layout-id"));
 
 		long oldLayoutId = layoutId;
